Store immediate timings in double (Fixes GCC 4.8+ compile error)
Former-commit-id: 4581381757e7243fef403250a21bf853ee628ef2

+++ mlp/test_gel_mlp.cpp
@@ -18,7 +18,7 @@ const int processed_frames = 100;
 int main()
 {
     carp::conductor_t conductor;
-    std::chrono::duration<float> elapsed_time(0);
+    std::chrono::duration<double> elapsed_time(0);
 
     int numberOfThreads =  tbb::tbb_thread::hardware_concurrency();
     PRINT(numberOfThreads);
